+++ 7.3.0/src/TakeyourStand.EntityFramework/Migrations/202012180651436_Upgraded_To_ABP_6_1_1.cs
@@ -9,8 +9,8 @@
         {
             DropPrimaryKey("dbo.AbpDynamicPropertyValues");
             DropPrimaryKey("dbo.AbpDynamicEntityPropertyValues");
-            AlterColumn("dbo.AbpDynamicPropertyValues", "Id", c => c.Long(nullable: false, identity: true));
-            AlterColumn("dbo.AbpDynamicEntityPropertyValues", "Id", c => c.Long(nullable: false, identity: true));
+            AlterColumn("dbo.AbpDynamicPropertyValues", "Id", c => c.Int(nullable: false, identity: true));
+            AlterColumn("dbo.AbpDynamicEntityPropertyValues", "Id", c => c.Int(nullable: false, identity: true));
             AddPrimaryKey("dbo.AbpDynamicPropertyValues", "Id");
             AddPrimaryKey("dbo.AbpDynamicEntityPropertyValues", "Id");
         }
